+++ include/stream/net_stream_audio_out.h
@@ -1,38 +1,3 @@
-//#ifndef NET_STREAM_AUDIO_OUT_HEADER
-//#define NET_STREAM_AUDIO_OUT_HEADER
-//
-//
-//#include "stream.h"
-//#include <wsock/udp_socket.h>
-//#include <pack/audio_packet.h>
-//
-//
-//namespace stream
-//{
-//	class NetStreamAudioOut : public Stream
-//	{
-//	private:
-//		pack::AudioPacket packet;
-//		wsock::udpSocket& sock;
-//		wsock::addr& remote;
-//		char sendbuf[2048];
-//		long long last = 0;
-//
-//
-//	public:
-//		NetStreamAudioOut(wsock::udpSocket& local, wsock::addr& remote_addr, int32_t ms);
-//		~NetStreamAudioOut();
-//
-//	public:
-//		virtual size_t stream_write(const void* frombuffer, size_t writesize) override;
-//
-//	private:
-//		virtual size_t stream_read(void* tobuffer, size_t buffersize, size_t readamount) override;
-//
-//	};
-//}
-//
-//#endif
 
 #ifndef NET_STREAM_AUDIO_OUT_HEADER
 #define NET_STREAM_AUDIO_OUT_HEADER
@@ -63,8 +28,8 @@ namespace stream
 		~NetStreamAudioOut();
 
 	public:
-		virtual size_t stream_write(const void* frombuffer, size_t writesize) override;
-		virtual size_t stream_read(void* tobuffer, size_t buffersize, size_t readamount) override;
+		virtual int64_t stream_write(const void* frombuffer, int64_t writesize) override;
+		virtual int64_t stream_read(void* tobuffer, int64_t buffersize, int64_t readamount) override;
 
 		void set_jitter_buffer(jbuf::JitterBuffer* jbuffer);
